Adding test for vec2 and remove a lot of conversion warnings

+++ game/include/game.h
@@ -1,14 +1,16 @@
 #ifndef GAME_H
 #define GAME_H
 
+
+#include <cstdlib>
 namespace game {
 
 struct LaunchOptions {
-  int initial_spawn_count = 0;
+  size_t initial_spawn_count = 0;
   int window_width = 1280;
   int window_height = 960;
-  int tilemap_width = 20;
-  int tilemap_height = 15;
+  size_t tilemap_width = 20;
+  size_t tilemap_height = 15;
 };
 
 void Loop(const LaunchOptions& options);
